+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.0.0 - TBD
+
+### Added
+
+- Nothing.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- [#7](https://github.com/zendframework/zend-expressive-helpers/pull/7) removes
+  the `RouteResultObserverInterface` implementation from `UrlHelper`. This also
+  means that the `UrlHelperMiddleware` no longer registers the `UrlHelper` as a
+  route result observer, but instead just injects it with the `RouteResult`
+  present as a request attribute, if any.
+
+### Fixed
+
+- Nothing.
+
 ## 1.4.1 - TBD
 
 ### Added